Add files via upload
In this particular project, we are using a dataset that contains information 
like State, account length, international plan, Total day cells, Total day 
charge and using that to predict whether the customer will remain a 
customer in the future.
However, before you go ahead and make a prediction, it is advised that 
you first pre-process the data, since it may contain some irregularities and 
noise. 
In addition, try various tricks and techniques in order to gain the best 
accuracy in your predictions.